I have created a function to check game state
If a player wins of it's a tie, you will be show the
result and the current score. There are still a few
bugs that I'll document later, and a lot of code to
refactor, but its 'working'. You can reset the display
to select a game mode, or play another round.