I have gotten the miniMax function working!
I made it so it returns the path and score in a
single object. I got over the deep copy and shallow
copy problems I was having because I was thinking in
C++ terms. Now all I need to do is change the grid
elements to reflect the choice of the ai.