I'm beginning to transition from divs to a canvas
The plan is to maintain most previous functions as
reference to what I need to be doing. Then I'll
create a game loop that will check for if I clicked
on the canvas. Then I'll implement the game's logic.
I think I'll still need to learn a lot more about the
canvas's properties, especially in relation to the
screen size.