Refactor all dom manipulation functions
I changed the program from using the DOM manipulation API to the basic
innerHTML property. I know that currently it runs slower because of string
parsing apparently, and also that it poses a security issue in some conditons
when the user provides an input that is not filtered and gets displayed,
but for this project in these conditions, I think that the code legibility
and the fact that it isn't a security risk, means its a good trade off.