Remove unecessary looping in runHumanVsHumanLoop
Before, I was using animationFrames (I don't know why) and that made
it so the checkGameState function kept running every frame. This also
meant that every frame, the canvas.onclick function was being reasigned
to the player logic. Anyways, this was completely useless, hurt performance
and needed to be removed.